+++ src/ReaderScreen/styles/index.js
@@ -1,0 +1,4 @@
+import { styles } from "./styles";
+import commonStyle from "./commonStyle";
+
+export { styles, commonStyle };
