+++ src/ReaderScreen/utils/index.js
@@ -1,0 +1,3 @@
+import { fontColorForReader, fontSizeForReader, convertToParagraph, processData } from "./util";
+
+export { fontColorForReader, fontSizeForReader, convertToParagraph, processData };
